@@ -16,9 +16,6 @@
     ],
     "config": {
         "sort-packages": true,
-        "platform": {
-            "php": "7.3.33"
-        },
         "allow-plugins": {
             "cweagans\/composer-patches": true,
             "dealerdirect\/phpcodesniffer-composer-installer": true,
@@ -27,7 +24,7 @@
         }
     },
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4 | ^8",
         "psr\/log": "^1.1.3",
         "psr\/container": "^1",
         "wpdesk\/wp-view": "^2",
