@@ -15,17 +15,17 @@
     ],
     "config": {
         "platform": {
-            "php": "7.0"
+            "php": "7.4"
         }
     },
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.4",
         "wpdesk\/wp-basic-requirements": "^3.2.3",
         "wpdesk\/wp-builder": "^2.0.0",
         "wpdesk\/wp-wpdesk-tracker": "^3"
     },
     "require-dev": {
-        "phpunit\/phpunit": "<7",
+        "phpunit\/phpunit": "^7||^8||^9",
         "wp-coding-standards\/wpcs": "^0.14.1",
         "squizlabs\/php_codesniffer": "^3.0.2",
         "mockery\/mockery": "*",
