@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=7.1 || ^8",
+        "php": ">=7.4 || ^8",
         "wpdesk\/wp-builder": "^2.0",
         "wpdesk\/wp-notice": "^3.1",
         "wpdesk\/wp-view": "^2",
