@@ -39,6 +39,8 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
         private $min_openssl_version = null;
         /** @var array */
         protected $plugin_require;
+        /** @var array */
+        protected $class_require;
         /** @var bool */
         protected $should_check_plugin_versions = \false;
         /** @var array */
@@ -73,6 +75,7 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
             $this->module_require = array();
             $this->setting_require = array();
             $this->notices = array();
+            $this->class_require = array();
         }
         /**
          * @param string $version
@@ -144,6 +147,20 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
             return $this;
         }
         /**
+         * Add a class to require list.
+         *
+         * @param string $class_name       Name of the class.
+         * @param string $plugin_nice_name Nice plugin name for better looks in notice.
+         *
+         * @return $this
+         *
+         */
+        public function add_class_require($class_name, $plugin_nice_name)
+        {
+            $this->class_require[$class_name] = $plugin_nice_name;
+            return $this;
+        }
+        /**
          * @param string $module_name
          * @param string $nice_name Nice module name for better looks in notice
          *
@@ -186,20 +203,21 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
         {
             $notices = array();
             if (!self::is_php_at_least($this->min_php_version)) {
-                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run on PHP versions older than %s. Please contact your host and ask them to upgrade.', $this->get_text_domain()), \esc_html($this->plugin_name), $this->min_php_version));
+                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run on PHP versions older than %s. Please contact your host and ask them to upgrade.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), $this->min_php_version));
             }
             if (!self::is_wp_at_least($this->min_wp_version)) {
-                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run on WordPress versions older than %s. Please update WordPress.', $this->get_text_domain()), \esc_html($this->plugin_name), $this->min_wp_version));
+                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run on WordPress versions older than %s. Please update WordPress.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), $this->min_wp_version));
             }
             if ($this->min_wc_version !== null && $this->can_check_plugin_version() && !self::is_wc_at_least($this->min_wc_version)) {
-                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run on WooCommerce versions older than %s. Please update WooCommerce.', $this->get_text_domain()), \esc_html($this->plugin_name), $this->min_wc_version));
+                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run on WooCommerce versions older than %s. Please update WooCommerce.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), $this->min_wc_version));
             }
             if ($this->min_openssl_version !== null && !self::is_open_ssl_at_least($this->min_openssl_version)) {
-                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without OpenSSL module version at least %s. Please update OpenSSL module.', $this->get_text_domain()), \esc_html($this->plugin_name), '0x' . \dechex($this->min_openssl_version)));
+                $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without OpenSSL module version at least %s. Please update OpenSSL module.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), '0x' . \dechex($this->min_openssl_version)));
             }
             $notices = $this->append_plugin_require_notices($notices);
             $notices = $this->append_module_require_notices($notices);
             $notices = $this->append_settings_require_notices($notices);
+            $notices = $this->append_class_require_notices($notices);
             if ($this->should_check_plugin_versions) {
                 $notices = $this->check_minimum_require_plugins_version_and_append_notices($notices);
             }
@@ -236,7 +254,11 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
          */
         public static function is_wp_at_least($min_version)
         {
-            return \version_compare(\get_bloginfo('version'), $min_version, '>=');
+            $wp_version = \get_bloginfo('version');
+            if ($wp_version === null || $wp_version === '') {
+                return \false;
+            }
+            return \version_compare($wp_version, $min_version, '>=');
         }
         /**
          * Are plugins loaded so we can check the version
@@ -256,7 +278,10 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
          */
         public static function is_wc_at_least($min_version)
         {
-            return \defined('WC_VERSION') && \version_compare(\WC_VERSION, $min_version, '>=');
+            if (!\defined('WC_VERSION') || \WC_VERSION === null || \WC_VERSION === '') {
+                return \false;
+            }
+            return \version_compare(\WC_VERSION, $min_version, '>=');
         }
         /**
          * Checks if ssl version is valid
@@ -281,8 +306,8 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
             $required_plugins = $this->retrieve_required_plugins_data();
             if (\count($required_plugins) > 0) {
                 foreach ($required_plugins as $plugin) {
-                    if (isset($plugin['Version']) && \version_compare($plugin['Version'], $plugin[self::PLUGIN_INFO_APPEND_PLUGIN_DATA], '<=')) {
-                        $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%1$s&#8221; plugin requires at least %2$s version of %3$s to work correctly. Please update it to its latest release.', $this->get_text_domain()), \esc_html($this->plugin_name), $plugin[self::PLUGIN_INFO_APPEND_PLUGIN_DATA], $plugin['Name']));
+                    if (isset($plugin['Version']) && $plugin['Version'] !== null && $plugin['Version'] !== '' && \version_compare($plugin['Version'], $plugin[self::PLUGIN_INFO_APPEND_PLUGIN_DATA], '<')) {
+                        $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%1$s&#8221; plugin requires at least %2$s version of %3$s to work correctly. Please update it to its latest release.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), $plugin[self::PLUGIN_INFO_APPEND_PLUGIN_DATA], $plugin['Name']));
                     }
                 }
             }
@@ -388,12 +413,27 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
                     if (isset($plugin_info['repository_url'])) {
                         $notice = $this->prepare_plugin_repository_require_notice($plugin_info);
                     } elseif (!self::is_wp_plugin_active($plugin_name)) {
-                        $notice = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.', $this->get_text_domain()), \esc_html($this->plugin_name), \esc_html(\basename($plugin_info[self::PLUGIN_INFO_KEY_NICE_NAME])), \esc_html(\basename($plugin_info[self::PLUGIN_INFO_KEY_NICE_NAME]))));
+                        $notice = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), \esc_html(\basename($plugin_info[self::PLUGIN_INFO_KEY_NICE_NAME])), \esc_html(\basename($plugin_info[self::PLUGIN_INFO_KEY_NICE_NAME]))));
                     }
                     if ($notice !== null) {
                         $notices[] = $notice;
                     }
                 }
+            }
+            return $notices;
+        }
+        /**
+         * @param array $notices
+         *
+         * @return array
+         */
+        private function append_class_require_notices($notices)
+        {
+            foreach ($this->class_require as $class_name => $plugin_nice_name) {
+                if (\class_exists($class_name)) {
+                    continue;
+                }
+                $notices[] = $this->prepare_notice_message(\sprintf('The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.', \esc_html($this->plugin_name), \esc_html($plugin_nice_name), \esc_html($plugin_nice_name)));
             }
             return $notices;
         }
@@ -406,23 +446,11 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
          */
         private function prepare_plugin_repository_install_url($plugin_info)
         {
-            $slug = \basename($plugin_info[self::PLUGIN_INFO_KEY_NAME]);
+            $slug = \basename(\explode('/', $plugin_info[self::PLUGIN_INFO_KEY_NAME])[0]);
             $install_url = \self_admin_url('update.php?action=install-plugin&plugin=' . $slug);
             if (\function_exists('wp_nonce_url') && \function_exists('wp_create_nonce')) {
                 $install_url = \wp_nonce_url($install_url, 'install-plugin_' . $slug);
             }
-            \add_filter('plugins_api', function ($api, $action, $args) use ($plugin_info, $slug) {
-                if ('plugin_information' !== $action || \false !== $api || !isset($args->slug) || $slug !== $args->slug) {
-                    return $api;
-                }
-                $api = new \stdClass();
-                $api->name = $plugin_info['nice_name'];
-                // self in closures requires 5.4
-                $api->version = '';
-                $api->download_link = \esc_url($plugin_info['repository_url']);
-                // self in closures requires 5.4
-                return $api;
-            }, 10, 3);
             return $install_url;
         }
         /**
@@ -437,13 +465,13 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
             if (!self::is_wp_plugin_active($name)) {
                 if (!self::is_wp_plugin_installed($name, $this->use_transients)) {
                     $install_url = $this->prepare_plugin_repository_install_url($plugin_info);
-                    return $this->prepare_notice_message(\sprintf(\wp_kses(\__('The &#8220;%s&#8221; plugin requires free %s plugin. <a href="%s">Install %s</a>', $this->get_text_domain()), array('a' => array('href' => array()))), $this->plugin_name, $nice_name, \esc_url($install_url), $nice_name));
+                    return $this->prepare_notice_message(\sprintf(\wp_kses(\__('The &#8220;%s&#8221; plugin requires free %s plugin. <a href="%s">Install %s</a>', 'custom-price-for-woocommerce'), array('a' => array('href' => array()))), $this->plugin_name, $nice_name, \esc_url($install_url), $nice_name));
                 }
                 $activate_url = 'plugins.php?action=activate&plugin=' . \urlencode($plugin_info[self::PLUGIN_INFO_KEY_NAME]) . '&plugin_status=all&paged=1&s';
                 if (\function_exists('wp_create_nonce')) {
                     $activate_url .= '&_wpnonce=' . \urlencode(\wp_create_nonce('activate-plugin_' . $name));
                 }
-                return $this->prepare_notice_message(\sprintf(\wp_kses(\__('The &#8220;%s&#8221; plugin requires activating %s plugin. <a href="%s">Activate %s</a>', $this->get_text_domain()), array('a' => array('href' => array()))), $this->plugin_name, $nice_name, \esc_url(\admin_url($activate_url)), $nice_name));
+                return $this->prepare_notice_message(\sprintf(\wp_kses(\__('The &#8220;%s&#8221; plugin requires activating %s plugin. <a href="%s">Activate %s</a>', 'custom-price-for-woocommerce'), array('a' => array('href' => array()))), $this->plugin_name, $nice_name, \esc_url(\admin_url($activate_url)), $nice_name));
             }
             return null;
         }
@@ -485,7 +513,7 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
             if (\count($this->module_require) > 0) {
                 foreach ($this->module_require as $module_name => $nice_module_name) {
                     if (!self::is_module_active($module_name)) {
-                        $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without %s PHP module installed. Please contact your host and ask them to install %s.', $this->get_text_domain()), \esc_html($this->plugin_name), \esc_html(\basename($nice_module_name)), \esc_html(\basename($nice_module_name))));
+                        $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without %s PHP module installed. Please contact your host and ask them to install %s.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), \esc_html(\basename($nice_module_name)), \esc_html(\basename($nice_module_name))));
                     }
                 }
             }
@@ -510,7 +538,7 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
             if (\count($this->setting_require) > 0) {
                 foreach ($this->setting_require as $setting => $value) {
                     if (!self::is_setting_set($setting, $value)) {
-                        $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without %s PHP setting set to %s. Please contact your host and ask them to set %s.', $this->get_text_domain()), \esc_html($this->plugin_name), \esc_html(\basename($setting)), \esc_html(\basename($value)), \esc_html(\basename($setting))));
+                        $notices[] = $this->prepare_notice_message(\sprintf(\__('The &#8220;%s&#8221; plugin cannot run without %s PHP setting set to %s. Please contact your host and ask them to set %s.', 'custom-price-for-woocommerce'), \esc_html($this->plugin_name), \esc_html(\basename($setting)), \esc_html(\basename($value)), \esc_html(\basename($setting))));
                     }
                 }
             }
@@ -595,7 +623,7 @@ if (!\class_exists('CPWFreeVendor\WPDesk_Basic_Requirement_Checker')) {
         public function handle_render_notices_action()
         {
             foreach ($this->notices as $notice) {
-                echo $notice;
+                echo \wp_kses_post($notice);
             }
         }
     }
