@@ -101,7 +101,7 @@ final class PluginBootstrap
     {
         add_action('activated_plugin', static function ($plugin_file, $network_wide = \false) {
             if (!$network_wide) {
-                $option_name = 'plugin_activation_' . $plugin_file;
+                $option_name = 'activation_plugin_' . $plugin_file;
                 $activation_date = get_option($option_name, '');
                 if ('' === $activation_date) {
                     $activation_date = current_time('mysql');
@@ -119,7 +119,14 @@ final class PluginBootstrap
         if (method_exists($plugin_info, 'get_language_dir')) {
             $lang_dir = $plugin_info->get_language_dir();
         }
-        \load_plugin_textdomain($plugin_info->get_text_domain(), \false, basename($plugin_info->get_plugin_dir()) . "/{$lang_dir}/");
+        $text_domain = $plugin_info->get_text_domain();
+        add_filter('doing_it_wrong_trigger_error', function ($doing_it_wrong, $function, $message, $version) use ($text_domain) {
+            if (wp_get_environment_type() === 'production' && $function === '_load_textdomain_just_in_time' && strpos($message, '<code>' . $text_domain . '</code>') !== \false) {
+                return \false;
+            }
+            return $doing_it_wrong;
+        }, 10, 4);
+        \load_plugin_textdomain($plugin_info->get_text_domain(), '', basename($plugin_info->get_plugin_dir()) . "/{$lang_dir}/");
     }
     /**
      * Factory method creates requirement checker to run the checks
@@ -149,7 +156,7 @@ final class PluginBootstrap
         $plugin_info->set_version($this->plugin_version);
         $plugin_info->set_product_id($this->product_id);
         $plugin_info->set_text_domain($this->plugin_text_domain);
-        $plugin_info->set_plugin_url(plugins_url(dirname(plugin_basename($this->plugin_file))));
+        $plugin_info->set_plugin_url(plugins_url('', $this->plugin_file));
         $plugin_info->set_plugin_shops($this->plugin_shops);
         return $plugin_info;
     }
